Create an abstract username/password strategy
